On second thought.. rawgit is a poor choice.
Rawgit caches permanently on the first access, which makes it impossible
to do bugfixes to bm.js.  We could use a specific commit hash or tag,
but then that would be built into the bookmark, and we'd still need to
update it.

Revert to using https://yerk.org/graphfi/bm.js, which we can redirect to
whatever is appropriate.